Moves the raw encryption message infrastructure to its own folder and namespace
